backend/requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.3
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-rest-framework==0.1.0
djangorestframework==3.11.0
Faker==4.0.3
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.9
mypy==0.770
mypy-extensions==0.4.3
mysqlclient==1.4.6
Pillow==7.1.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-jsonrpc-server==0.3.4
pytz==2019.3
redis==3.5.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
uzu-accounts-app==1.1.9
vine==1.3.0
zipp==3.1.0
